Added security and use cases on core module.

+++ review-manager-core/src/main/java/com/tracktainment/reviewmanager/dto/ReviewUpdate.java
@@ -22,7 +22,7 @@ public class ReviewUpdate {
     @Min(value = 1, message = Constants.RATING_INVALID_MSG)
     @Max(value = 5, message = Constants.RATING_INVALID_MSG)
     @Schema(description = "Rating of the reviewed entity", example = "5")
-    private int rating;
+    private float rating;
 
     @Pattern(regexp = Constants.TITLE_REGEX, message = Constants.TITLE_INVALID_MSG)
     @Schema(description = "Title of the review", example = "Captivating story with rich character development")
